[wip][mlir][Vector] Update the lowering patterns for vector.contract
Instead of "unrolling" and generating a sequence of
`vector.outerproduct` Ops (which will not work for scalable vectors),
generate an `scf.for` loops when lowering `vector.contract` to
`vector.outerproduct`.

See the update test for "before" and "after" comparison.

TODO:
Make sure that `vector.extract` can be lowered to LLVM when using
dynamic indices. That would be a seperate bit of work.